Split out activate
