+++ client/state/themes/actions/index.js
@@ -25,21 +25,10 @@ import {
 	THEME_PREVIEW_STATE,
 } from 'state/action-types';
 import { recordTracksEvent, withAnalytics } from 'state/analytics/actions';
-import {
-	getTheme,
-	themeHasAutoLoadingHomepage,
-	hasAutoLoadingHomepageModalAccepted,
-} from 'state/themes/selectors';
-import { isJetpackSite } from 'state/sites/selectors';
 
 import 'state/data-layer/wpcom/theme-filters';
 
 import 'state/themes/init';
-
-import { activateTheme } from 'state/themes/actions/activate-theme';
-import { suffixThemeIdForInstall } from 'state/themes/actions/suffix-theme-id-for-install';
-import { installAndActivateTheme } from 'state/themes/actions/install-and-activate-theme';
-import { showAutoLoadingHomepageWarning } from 'state/themes/actions/show-auto-loading-homepage-warning';
 
 export { setBackPath } from 'state/themes/actions/set-back-path';
 export { receiveThemes } from 'state/themes/actions/receive-themes';
@@ -61,41 +50,7 @@ export { clearThemeUpload } from 'state/themes/actions/clear-theme-upload';
 export { deleteTheme } from 'state/themes/actions/delete-theme';
 export { confirmDelete } from 'state/themes/actions/confirm-delete';
 export { showAutoLoadingHomepageWarning } from 'state/themes/actions/show-auto-loading-homepage-warning';
-
-/**
- * Triggers a network request to activate a specific theme on a given site.
- * If it's a Jetpack site, installs the theme prior to activation if it isn't already.
- *
- * @param  {string}   themeId   Theme ID
- * @param  {number}   siteId    Site ID
- * @param  {string}   source    The source that is requesting theme activation, e.g. 'showcase'
- * @param  {boolean}  purchased Whether the theme has been purchased prior to activation
- * @returns {Function}          Action thunk
- */
-export function activate( themeId, siteId, source = 'unknown', purchased = false ) {
-	return ( dispatch, getState ) => {
-		/**
-		 * Let's check if the theme will change the homepage of the site,
-		 * before to definitely start the theme-activating process,
-		 * allowing cancel it if it's desired.
-		 */
-		if (
-			themeHasAutoLoadingHomepage( getState(), themeId ) &&
-			! hasAutoLoadingHomepageModalAccepted( getState(), themeId )
-		) {
-			return dispatch( showAutoLoadingHomepageWarning( themeId ) );
-		}
-
-		if ( isJetpackSite( getState(), siteId ) && ! getTheme( getState(), siteId, themeId ) ) {
-			const installId = suffixThemeIdForInstall( getState(), siteId, themeId );
-			// If theme is already installed, installation will silently fail,
-			// and it will just be activated.
-			return dispatch( installAndActivateTheme( installId, siteId, source, purchased ) );
-		}
-
-		return dispatch( activateTheme( themeId, siteId, source, purchased ) );
-	};
-}
+export { activate } from 'state/themes/actions/activate';
 
 /**
  * Start an Automated Transfer with an uploaded theme.
